Update patch set 1
Patch Set 1:

I'm not sure if we should do this, as CDMA is mostly working (apart from LTE), there are a large number of existing CDMA users, and many of them are already making their LTE work by changing APNs. I would just ship it as is for now to avoid cutting off existing users (of my unofficial builds), and eventually automate the APN replacement in a manner similar to osprey.

Patch-set: 1