Update patch set 1
Patch Set 1:

I understand your concern about less technically savvy users. Your idea of refusing to install and issuing a warning when CDMA and the fsg-id is not set would be a better approach. That way, experienced users can continue to use it while we avoid a flood of complaints from new users.

Patch-set: 1