Update patch set 2
Patch Set 2:

(3 comments)

See the comments and topic is unnecessary for just a patch. The best thing to do here is rename it to rebrand-moto8916 and add https://review.lineageos.org/#/c/154524/ to it. Also osprey, lux and harpia related changes should be under the same topic of course.

Patch-set: 2
Label: Verified=0